script: enable continous search mode
In normal operation, rofi executes the script at startup to gather
the displayed information. Nevertheless, for some inputs it is useful to
have an continous mode (e.g., desktop search). In continous mode, the
underlying script is executed for every keystroke to update the results.

In this patch the script syntax is extended to capture an option to a
script dialog:

  -modi "file:recoll.sh:c"

The ':c' enables the continous mode for the dialog.