Initialize locale upon startup.
It seems that Xlib input contexts assume a C locale when the locale
hasn’t been initialized from the environment before the input context
is created. Inter alia, this will lead to Xlib reading the .XCompose
definition file as ISO-8859-1, which will result in mojibake when
composing characters in a UTF-8 environment (cf. #268).