Match X11 dpi by default
If no specific dpi is requested match the one of the X11
screen we connected to by default.
This is necessary for out-of-the-box support of HIDPI monitors,
especially when the user - e.g. on a laptop with such
an internal monitors - frequently switches between internal
and external (normal dpi) screens.