Fix rofi-sensible-terminal for termite.
When using the '-e' flag, termite expects that the program to be run and
all its arguments will be passed as a single string argument, which is
unlike some other terminals. Because of this, rofi-sensible-terminal was
broken for termite, e.g. trying to connect to a host using rofi -show ssh
doesn't open a terminal successfully.