Fix -m -4 for xmonad
The problem is that xcb_translate_coordinates was being called to
translate the origin of the active_window (relative to active_window's
parent) from the coordinate space relative to active_window to the
coordinate space relative to root. The fix is to translate the origin of
the active_window (relative to active_window's parent) from the
coordinate space relative to active_window's parent to the coordinate
space relative to root.