Add new "start" matching method
This is similar to "normal", except that it will only match the *start* of
strings.

For example, using `rofi -show run -matching normal` and then typing `pid` on my
system shows (in this order):

	acpid
	pidgin
	pidof

But with `-matching start` it's:

	pidgin
	pidof

Many more examples where the output is much more useful, IMHO.

Could also use `matching regex` and `^pid`, but that's one extra character of
typing I need to do for ever `rofi` invocation, *and* I'd then have to worry
about escaping regexps, which I'd rather not do.