xrmoptions: add element-height alias for eh
When looking through other people's Rofi configuration, one might
stumble upon the option `eh`.  It's hard to tell what it does, and it is
not found in the man pages.  Create an alias `element-height` for `eh`
that is more understandable.  Hopefully, people start adopting this in
their config files for readability.

EDIT: it has been added to the documentation recently, commits
5b8aebad3e9be5cf22f2c313e58dec1b56ef7d9a and
aa07b8ef9475032634492f42a4f568c01fc47763.